+++ src/settings/components/Sets/SetsForm/components/FilteringConditions/components/FilteringConditionsName/FilteringConditionsName.js
@@ -1,6 +1,6 @@
-import { FormattedMessage } from 'react-intl';
-import PropTypes from 'prop-types';
 import React from 'react';
+import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
 
 const FilteringConditionsName = ({
   input: {