+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Add form field validation. Refs UIOAIPMH-12.
 * Show set details next to the list of sets. Refs UIOAIPMH-25.
 * Update stripes to v5. Refs UIOAIPMH-28.
+* Add set spec specification on create/edit page. Refs UIOAIPMH-21.
 
 ## [1.2.0](https://github.com/folio-org/ui-oai-pmh/tree/v1.2.0) (2020-07-03)
 [Full Changelog](https://github.com/folio-org/ui-oai-pmh/compare/v1.1.0...v1.2.0)